Enemies Updates + Potion Mechanic (No Visuals Only Core Mechanic)

+++ Project Smash/Assets/_Scripts/Attributes/Ground.cs
@@ -16,7 +16,7 @@ namespace PSmash.Attributes
             audioSource.clip = damageObjectSound;
             audioSource.playOnAwake = false;
         }
-        public void TakeDamage(Transform attacker, int damage)
+        public void TakeDamage(Transform attacker, float damage)
         {
             audioSource.pitch = Random.Range(0.7f, 1);
             audioSource.Play();
